# src/main/resources/openapi/open-api.yaml
openapi: 3.0.3
info:
  title: Rock, Scissors, Paper Game API
  description: API for playing Rock, Scissors, and Paper game
  version: 1.0.0
  contact:
    email: david.martinez.dev@gmail.com

servers:
  - url: /api/game

paths:
  /play:
    post:
      summary: Play a round of Rock, Scissors, Paper
      operationId: playGame
      tags:
        - Game
      requestBody:
        description: User's choice in the game
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Game"
      responses:
        200:
          description: Game result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GameResult"
  /games:
    get:
      summary: Get games based on a nickname
      operationId: getGames
      tags:
        - Game
      parameters:
        - $ref: "#/components/parameters/nickname"
      responses:
        200:
          description: Statistics for nickname
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserStats"

components:
  parameters:
    nickname:
      name: nickname
      in: query
      required: true
      description: nickname from user
      schema:
        type: string
  schemas:
    GameResult:
      type: object
      properties:
        nickname:
          type: string
        user_choice:
          $ref: "#/components/schemas/Choice"
        machine_choice:
          $ref: "#/components/schemas/Choice"
        result:
          $ref: "#/components/schemas/Result"
      required:
        - user_choice
        - machine_choice
        - result
    Game:
      type: object
      properties:
        nickname:
          type: string
        choice:
          $ref: "#/components/schemas/Choice"
      required:
        - nickname
        - choice
    Choice:
      description: The user's choice in the game.
      type: string
      enum:
        - ROCK
        - SCISSORS
        - PAPER
    Result:
      description: The result of the game for the user.
      type: string
      enum:
        - WIN
        - LOSE
        - DRAW
    UserStats:
      type: object
      properties:
        nickname:
          type: string
        wins:
          type: number
        draws:
          type: number
        losses:
          type: number